Add "auto compute" check box
When checked, nothing change.

When unchecked, disable phraseChangeTimeoutEvent() so there is no update. Also, add an error message "Auto compute is disabled" so after changing phrase/passephrase and parameters, you are reminded to check "auto compute" to compute your seed.

I added this because on very low end computer (raspberry pi) the computing is quite annoying and you have to wait for each change.

When you add a custom words, seed is computed every 400ms so you have to copy/paste it to not be stuck after each letter.

And if you encrypt your private key with BIP38 it become unusable.

Autocompute is by default checked so there is no change for other users.